+++ src/models/DemoFileCache.ts
@@ -1,0 +1,4 @@
+export interface DemoFileCache {
+  filePath: string;
+  demo: { idx: number; title: string }[];
+}