+++ src/models/index.ts
@@ -1,3 +1,4 @@
 export * from "./Action";
+export * from "./DemoFileCache";
 export * from "./Demos";
 export * from "./Subscription";